+++ py/prospect/scripts/specview_cmx_targets.py
@@ -31,6 +31,7 @@ def _parse():
     parser.add_argument('--titlepage_prefix', help='Prefix for webpage title', type=str, default='targetlist')
     parser.add_argument('--with_multiple_models', help='Display several models (requires full redrock outputs)', action='store_true')
     parser.add_argument('--template_dir', help='Redrock template directory', type=str, default=None)
+    parser.add_argument('--countdown', help='Insert a countdown widget (in minutes)', type=int, default=-1)
 
     args = parser.parse_args()
     return args
@@ -84,7 +85,7 @@ def main():
         plotspectra(thespec, with_noise=True, is_coadded=True, zcatalog=the_zcat,
                     title=titlepage, html_dir=args.webdir, mask_type='CMX_TARGET', with_thumb_only_page=True,
                     template_dir=args.template_dir, redrock_cat=the_rrtable, num_approx_fits=num_approx_fits,
-                    with_full_2ndfit=with_full_2ndfit)
+                    with_full_2ndfit=with_full_2ndfit, vi_countdown=args.countdown)
 
     log.info("End of specview_cmx_targets script.")
     return 0